+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/configure/ReflectionConfigurationParser.java
@@ -83,9 +83,15 @@ public final class ReflectionConfigurationParser<C, T> extends ConfigurationPars
         if (!conditionResult.isPresent()) {
             return;
         }
+        C condition = conditionResult.get();
 
-        Optional<String> className = parseType(data);
-        if (className.isEmpty()) {
+        /*
+         * Classes registered using the old ("class") syntax will require elements (fields, methods,
+         * constructors, ...) to be registered for runtime queries, whereas the new ("type") syntax
+         * will automatically register all elements as queried.
+         */
+        Optional<ConfigurationTypeDescriptor> type = parseType(data);
+        if (type.isEmpty()) {
             return;
         }
 
@@ -93,10 +99,9 @@ public final class ReflectionConfigurationParser<C, T> extends ConfigurationPars
          * Even if primitives cannot be queried through Class.forName, they can be registered to
          * allow getDeclaredMethods() and similar bulk queries at run time.
          */
-        C condition = conditionResult.get();
-        TypeResult<T> result = delegate.resolveType(condition, className.get(), true, false);
+        TypeResult<T> result = delegate.resolveType(condition, type.get(), true, false);
         if (!result.isPresent()) {
-            handleMissingElement("Could not resolve class " + className.get() + " for reflection configuration.", result.getException());
+            handleMissingElement("Could not resolve " + type.get() + " for reflection configuration.", result.getException());
             return;
         }
         T clazz = result.get();
@@ -281,7 +286,7 @@ public final class ReflectionConfigurationParser<C, T> extends ConfigurationPars
         List<T> result = new ArrayList<>();
         for (Object type : types) {
             String typeName = asString(type, "types");
-            TypeResult<T> typeResult = delegate.resolveType(conditionResolver.alwaysTrue(), typeName, true, false);
+            TypeResult<T> typeResult = delegate.resolveType(conditionResolver.alwaysTrue(), new NamedConfigurationTypeDescriptor(typeName), true, false);
             if (!typeResult.isPresent()) {
                 handleMissingElement("Could not register method " + formatMethod(clazz, methodName) + " for reflection.", typeResult.getException());
                 return null;